+++ protocol/test/contracts/TestERC20.sol
@@ -4,11 +4,11 @@ pragma solidity ^0.8.30;
 import "@openzeppelin/contracts-upgradeable/token/ERC20/ERC20Upgradeable.sol";
 
 contract TestERC20 is ERC20Upgradeable {
-    function mint(address to, uint amount) external {
+    function mint(address to, uint256 amount) external {
         _mint(to, amount);
     }
 
-    function mint(uint amount) external {
+    function mint(uint256 amount) external {
         _mint(_msgSender(), amount);
     }
