+++ protocol/test/contracts/dependencies/RenderingContract.sol
@@ -2,7 +2,7 @@
 pragma solidity ^0.8.30;
 
 interface ITokenURIGenerator {
-    function tokenURI(uint tokenId) external view returns (string memory);
+    function tokenURI(uint256 tokenId) external view returns (string memory);
 }
 
 abstract contract RenderingContract {
